+++ src/main/java/cart/dao/MemberDao.java
@@ -31,8 +31,8 @@ public class MemberDao {
                 .build();
     }
 
-    public int findIdByAuthInfo(final String email, final String password) {
+    public Long findIdByAuthInfo(final String email, final String password) {
         final String query = "SELECT id FROM MEMBER WHERE email = ? AND password = ?";
-        return jdbcTemplate.queryForObject(query, Integer.class, email, password);
+        return jdbcTemplate.queryForObject(query, Long.class, email, password);
     }
 }
